(#3) Replaced noise and RNG with interfaces.
