+++ Assets/Application/Interfaces/IHeightSource.cs
@@ -1,0 +1,7 @@
+namespace Application.Interfaces
+{
+    public interface IHeightSource
+    {
+        float GetUnitHeight(int x, int y);
+    }
+}
