+++ Assets/Infrastructure/Runtime/SeedService.cs
@@ -4,7 +4,7 @@ namespace Infrastructure.Runtime
 {
     public class SeedService : ISeedService
     {
-        public int Seed { get; private set; } = 1;
+        public int Seed { get; private set; } = 1128;
 
         public void SetSeed(int seed)
             => Seed = seed;
