--- Assets/Scripts/PerlinNoise.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8aff354a9d93443e912358f6be1b0855
-timeCreated: 1649561714
