--- Assets/Scripts/PseudoRandomGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb911364278299b479e9dbd0d6e4ab64
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
